extract pm quality policy routes
